Add files via upload
The K-means clustering algorithm is a technique that can be used to group together similar items in a dataset. In the context of the mall customers dataset, this means that the algorithm can be used to group together customers who share similar characteristics. This can be useful for marketing purposes, as it can help to identify groups of customers who are more likely to respond positively to certain types of marketing campaigns.